Изменил работу с Money (бывший Sum) и ReceiptFiscalData и сделал их нормальными Value Object.

+++ ClientServerContracts/Budgets/GetBudgetDetail/GetBudgetsDetailResponse.cs
@@ -2,4 +2,4 @@ using ClientServerContracts.Budgets.GetBudgetSpendings;
 
 namespace ClientServerContracts.Budgets.GetBudgetDetail;
 
-public sealed record  GetBudgetsDetailResponse(Guid Id,  string Name,  int BeginOfPeriod, int? Limit);
+public sealed record  GetBudgetsDetailResponse(Guid Id,  string Name,  int BeginOfPeriod, long? Limit);
