+++ ClientServerContracts/Budgets/GetBudgetSpendings/SpendingDto.cs
@@ -1,3 +1,3 @@
 namespace ClientServerContracts.Budgets.GetBudgetSpendings;
 
-public sealed record SpendingDto(Guid Id, string Description, int Sum);
+public sealed record SpendingDto(Guid Id, string Description, long Sum);
